+++ src/db/mysql/sociConversions.h
@@ -163,7 +163,7 @@ namespace soci
             job.priority   = v.get<int>("priority");
             job.voName     = v.get<std::string>("vo_name");
             // COUNT(*) type is long long inside soci
-            job.numFiles   = v.get<long long>("numFiles");
+            job.numFiles   = v.get<int>("numFiles");
 
             try {
                 job.fileStatus = v.get<std::string>("file_state");
